Move StoryReorganizer into webarchive package
