--- src/main/scala/whodarr/classic/organizer/SeasonReorganizer.scala
@@ -1,1 +1,0 @@
-
