+++ src/main/scala/whodarr/classic/organizer/webarchive/SeasonReorganizer.scala
@@ -1,0 +1,1 @@
+package whodarr.classic.organizer.webarchive
